+++ libs/utils/sort.c
@@ -1,6 +1,6 @@
 #include <utils.h>
 
-static void compare_elements(long *a, long *b) {
+static void compare_elements(int *a, int *b) {
 	if (*b > *a) {
 		long temp = *a;
 		*a = *b;
@@ -8,7 +8,7 @@ static void compare_elements(long *a, long *b) {
 	}
 }
 
-void sort(long *data, const unsigned long size) {
+void sort(int *data, const unsigned long size) {
 	for (unsigned long a = 0; a < size - 1; ++a) {
 		for (unsigned long i = 0; i < size - a - 1; ++i) {
 			compare_elements(&data[i], &data[i + 1]);
